Change the class name and file name of filedisplaypreparation to ls and file name of ls to main
より適切なクラス名称とするために、filedisplaypreparation のクラス名とファイル名を ls に、ls のファイル名を main に変更
