+++ 07.ls_object/main.rb
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require_relative 'ls'
+
+Ls.new.display_files_based_on_format
